+++ java/client/src/main/java/module-info.java
@@ -6,10 +6,10 @@ module glide.api {
     exports glide.api.models.commands.bitmap;
     exports glide.api.models.commands.geospatial;
     exports glide.api.models.commands.function;
+    exports glide.api.models.commands.scan;
     exports glide.api.models.commands.stream;
     exports glide.api.models.configuration;
     exports glide.api.models.exceptions;
-    exports glide.api.models.commands.scan;
 
     requires com.google.protobuf;
     requires io.netty.codec;