+++ python/python/tests/test_async_client.py
@@ -6547,7 +6547,7 @@ class TestCommands:
         assert await redis_client.object_encoding(string_key) == "embstr"
 
         assert await redis_client.lpush(list_key, ["1"]) == 1
-        if await check_if_server_version_lt(redis_client, "7.0.0"):
+        if await check_if_server_version_lt(redis_client, "7.2.0"):
             assert await redis_client.object_encoding(list_key) == "quicklist"
         else:
             assert await redis_client.object_encoding(list_key) == "listpack"
@@ -6673,6 +6673,7 @@ class TestCommands:
         assert await redis_client.function_load(code, True) == lib_name
 
         func2_name = f"myfunc2c{get_random_string(5)}"
+        new_code = f"""{code}\n redis.register_function({func2_name}, function(keys, args) return #args end)"""
         new_code = generate_lua_lib_code(
             lib_name, {func_name: "return args[1]", func2_name: "return #args"}, True
         )
@@ -6725,6 +6726,7 @@ class TestCommands:
         assert await redis_client.function_load(code, True, route) == lib_name
 
         func2_name = f"myfunc2c{get_random_string(5)}"
+        new_code = f"""{code}\n redis.register_function({func2_name}, function(keys, args) return #args end)"""
         new_code = generate_lua_lib_code(
             lib_name, {func_name: "return args[1]", func2_name: "return #args"}, True
         )
@@ -6790,6 +6792,62 @@ class TestCommands:
 
         # Clean up by flushing functions again
         assert await redis_client.function_flush(route=route) == OK
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_function_delete(self, redis_client: TGlideClient):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(redis_client, min_version):
+            pytest.skip(f"Redis version required >= {min_version}")
+
+        lib_name = f"mylib1C{get_random_string(5)}"
+        func_name = f"myfunc1c{get_random_string(5)}"
+        code = generate_lua_lib_code(lib_name, {func_name: "return args[1]"}, True)
+
+        # Load the function
+        assert await redis_client.function_load(code) == lib_name
+
+        # TODO: Ensure the library exists with FUNCTION LIST
+
+        # Delete the function
+        assert await redis_client.function_delete(lib_name) == OK
+
+        # TODO: Ensure the function is no longer present with FUNCTION LIST
+
+        # deleting a non-existing library
+        with pytest.raises(RequestError) as e:
+            await redis_client.function_delete(lib_name)
+        assert "Library not found" in str(e)
+
+    @pytest.mark.parametrize("cluster_mode", [True])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    @pytest.mark.parametrize("single_route", [True, False])
+    async def test_function_delete_with_routing(
+        self, redis_client: GlideClusterClient, single_route: bool
+    ):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(redis_client, min_version):
+            pytest.skip(f"Redis version required >= {min_version}")
+
+        lib_name = f"mylib1C{get_random_string(5)}"
+        func_name = f"myfunc1c{get_random_string(5)}"
+        code = generate_lua_lib_code(lib_name, {func_name: "return args[1]"}, True)
+        route = SlotKeyRoute(SlotType.PRIMARY, "1") if single_route else AllPrimaries()
+
+        # Load the function
+        assert await redis_client.function_load(code, False, route) == lib_name
+
+        # TODO: Ensure the library exists with FUNCTION LIST
+
+        # Delete the function
+        assert await redis_client.function_delete(lib_name, route) == OK
+
+        # TODO: Ensure the function is no longer present with FUNCTION LIST
+
+        # deleting a non-existing library
+        with pytest.raises(RequestError) as e:
+            await redis_client.function_delete(lib_name)
+        assert "Library not found" in str(e)
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
@@ -7102,6 +7160,147 @@ class TestCommands:
         # DB 0 should still have no keys, so random_key should still return None
         assert await redis_client.random_key() is None
 
+    @pytest.mark.parametrize("cluster_mode", [False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_lcs(self, redis_client: GlideClient):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(redis_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+        key1 = "testKey1"
+        value1 = "abcd"
+        key2 = "testKey2"
+        value2 = "axcd"
+        nonexistent_key = "nonexistent_key"
+        expected_subsequence = "acd"
+        expected_subsequence_with_nonexistent_key = ""
+        assert await redis_client.mset({key1: value1, key2: value2}) == OK
+        assert await redis_client.lcs(key1, key2) == expected_subsequence
+        assert (
+            await redis_client.lcs(key1, nonexistent_key)
+            == expected_subsequence_with_nonexistent_key
+        )
+        lcs_non_string_key = "lcs_non_string_key"
+        assert await redis_client.sadd(lcs_non_string_key, ["Hello", "world"]) == 2
+        with pytest.raises(RequestError):
+            await redis_client.lcs(key1, lcs_non_string_key)
+
+    @pytest.mark.parametrize("cluster_mode", [False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_lcs_len(self, redis_client: GlideClient):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(redis_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+        key1 = "testKey1"
+        value1 = "abcd"
+        key2 = "testKey2"
+        value2 = "axcd"
+        nonexistent_key = "nonexistent_key"
+        expected_subsequence_length = 3
+        expected_subsequence_length_with_nonexistent_key = 0
+        assert await redis_client.mset({key1: value1, key2: value2}) == OK
+        assert await redis_client.lcs_len(key1, key2) == expected_subsequence_length
+        assert (
+            await redis_client.lcs_len(key1, nonexistent_key)
+            == expected_subsequence_length_with_nonexistent_key
+        )
+        lcs_non_string_key = "lcs_non_string_key"
+        assert await redis_client.sadd(lcs_non_string_key, ["Hello", "world"]) == 2
+        with pytest.raises(RequestError):
+            await redis_client.lcs_len(key1, lcs_non_string_key)
+
+    @pytest.mark.parametrize("cluster_mode", [False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_lcs_idx(self, redis_client: GlideClient):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(redis_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+        key1 = "testKey1"
+        value1 = "abcd1234"
+        key2 = "testKey2"
+        value2 = "bcdef1234"
+        nonexistent_key = "nonexistent_key"
+        expected_response_no_min_match_len_no_with_match_len = {
+            "matches": [
+                [
+                    [4, 7],
+                    [5, 8],
+                ],
+                [
+                    [1, 3],
+                    [0, 2],
+                ],
+            ],
+            "len": 7,
+        }
+        expected_response_with_min_match_len_equals_four_no_with_match_len = {
+            "matches": [
+                [
+                    [4, 7],
+                    [5, 8],
+                ],
+            ],
+            "len": 7,
+        }
+        expected_response_no_min_match_len_with_match_len = {
+            "matches": [
+                [
+                    [4, 7],
+                    [5, 8],
+                    4,
+                ],
+                [
+                    [1, 3],
+                    [0, 2],
+                    3,
+                ],
+            ],
+            "len": 7,
+        }
+        expected_response_with_min_match_len_equals_four_and_with_match_len = {
+            "matches": [
+                [
+                    [4, 7],
+                    [5, 8],
+                    4,
+                ],
+            ],
+            "len": 7,
+        }
+        expected_response_with_nonexistent_key = {
+            "matches": [],
+            "len": 0,
+        }
+        assert await redis_client.mset({key1: value1, key2: value2}) == OK
+        assert (
+            await redis_client.lcs_idx(key1, key2)
+            == expected_response_no_min_match_len_no_with_match_len
+        )
+        assert (
+            await redis_client.lcs_idx(key1, key2, min_match_len=4)
+            == expected_response_with_min_match_len_equals_four_no_with_match_len
+        )
+        assert (
+            # negative min_match_len should have no affect on the output
+            await redis_client.lcs_idx(key1, key2, min_match_len=-3)
+            == expected_response_no_min_match_len_no_with_match_len
+        )
+        assert (
+            await redis_client.lcs_idx(key1, key2, with_match_len=True)
+            == expected_response_no_min_match_len_with_match_len
+        )
+        assert (
+            await redis_client.lcs_idx(key1, key2, min_match_len=4, with_match_len=True)
+            == expected_response_with_min_match_len_equals_four_and_with_match_len
+        )
+        assert (
+            await redis_client.lcs_idx(key1, nonexistent_key)
+            == expected_response_with_nonexistent_key
+        )
+        lcs_non_string_key = "lcs_non_string_key"
+        assert await redis_client.sadd(lcs_non_string_key, ["Hello", "world"]) == 2
+        with pytest.raises(RequestError):
+            await redis_client.lcs_idx(key1, lcs_non_string_key)
+
 
 class TestMultiKeyCommandCrossSlot:
     @pytest.mark.parametrize("cluster_mode", [True])
@@ -7167,6 +7366,9 @@ class TestMultiKeyCommandCrossSlot:
                     redis_client.sintercard(["def", "ghi"]),
                     redis_client.lmpop(["def", "ghi"], ListDirection.LEFT),
                     redis_client.blmpop(["def", "ghi"], ListDirection.LEFT, 1),
+                    redis_client.lcs("abc", "def"),
+                    redis_client.lcs_len("abc", "def"),
+                    redis_client.lcs_idx("abc", "def"),
                 ]
             )
 
@@ -7461,6 +7663,90 @@ class TestClusterRoutes:
             assert await redis_client.dbsize() > 0
             assert await redis_client.flushdb(FlushMode.SYNC, AllPrimaries()) is OK
             assert await redis_client.dbsize() == 0
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_sscan(self, redis_client: GlideClusterClient):
+        key1 = f"{{key}}-1{get_random_string(5)}"
+        key2 = f"{{key}}-2{get_random_string(5)}"
+        initial_cursor = "0"
+        result_cursor_index = 0
+        result_collection_index = 1
+        default_count = 10
+        num_members = list(
+            map(str, range(50000))
+        )  # Use large dataset to force an iterative cursor.
+        char_members = ["a", "b", "c", "d", "e"]
+
+        # Empty set
+        result = await redis_client.sscan(key1, initial_cursor)
+        assert result[result_cursor_index] == initial_cursor
+        assert result[result_collection_index] == []
+
+        # Negative cursor
+        result = await redis_client.sscan(key1, "-1")
+        assert result[result_cursor_index] == initial_cursor
+        assert result[result_collection_index] == []
+
+        # Result contains the whole set
+        assert await redis_client.sadd(key1, char_members) == len(char_members)
+        result = await redis_client.sscan(key1, initial_cursor)
+        assert result[result_cursor_index] == initial_cursor
+        assert len(result[result_collection_index]) == len(char_members)
+        assert set(result[result_collection_index]).issubset(set(char_members))
+
+        result = await redis_client.sscan(key1, initial_cursor, match="a")
+        assert result[result_cursor_index] == initial_cursor
+        assert set(result[result_collection_index]).issubset(set(["a"]))
+
+        # Result contains a subset of the key
+        assert await redis_client.sadd(key1, num_members) == len(num_members)
+        result_cursor = "0"
+        result_values = set()  # type: set[str]
+        result = await redis_client.sscan(key1, result_cursor)
+        result_cursor = str(result[result_cursor_index])
+        result_values.update(result[result_collection_index])
+
+        # 0 is returned for the cursor of the last iteration.
+        while result_cursor != "0":
+            next_result = await redis_client.sscan(key1, result_cursor)
+            next_result_cursor = str(next_result[result_cursor_index])
+            assert next_result_cursor != result_cursor
+
+            assert not set(result[result_collection_index]).issubset(
+                set(next_result[result_collection_index])
+            )
+            result_values.update(next_result[result_collection_index])
+            result_cursor = next_result_cursor
+        assert set(num_members).issubset(result_values)
+        assert set(char_members).issubset(result_values)
+
+        # Test match pattern
+        result = await redis_client.sscan(key1, initial_cursor, match="*")
+        assert result[result_cursor_index] != "0"
+        assert len(result[result_collection_index]) >= default_count
+
+        # Test count
+        result = await redis_client.sscan(key1, initial_cursor, count=20)
+        assert result[result_cursor_index] != "0"
+        assert len(result[result_collection_index]) >= 20
+
+        # Test count with match returns a non-empty list
+        result = await redis_client.sscan(key1, initial_cursor, match="1*", count=20)
+        assert result[result_cursor_index] != "0"
+        assert len(result[result_collection_index]) >= 0
+
+        # Exceptions
+        # Non-set key
+        assert await redis_client.set(key2, "test") == OK
+        with pytest.raises(RequestError):
+            await redis_client.sscan(key2, initial_cursor)
+        with pytest.raises(RequestError):
+            await redis_client.sscan(key2, initial_cursor, match="test", count=20)
+
+        # Negative count
+        with pytest.raises(RequestError):
+            await redis_client.sscan(key2, initial_cursor, count=-1)
 
 
 @pytest.mark.asyncio